pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "berlinbnb"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.3.1",
    "pyspark>=4.0.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "selenium>=4.34.2",
    "tabulate>=0.9.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["*.py"]
exclude = ["tests*", "notebooks*", "terraform*", "data*", "sample_data*", "artifacts*"]
